Add files via upload
flask v1 of the app. Concerns include text parsing and subprocess module in order to use 2.7 tensorflow script with future development of 3.5 flask. Need to write tensorflow script in 3.5 or possibly use c++ to run the necessary modules. This is a change from the other flask version of the app that relies upon integrated tensorflow script limiting development to 2.7 which is not the direction I want to take this.